requirements.txt
python-dotenv
six
matplotlib
sympy
fastapi[all]
uvicorn[standard]
langchain
langchain-community
openpyxl
langchain_experimental
langchain-community
tiktoken
faiss-cpu
chromadb
PyPDF2
docx2txt
sentence-transformers
InstructorEmbedding
beautifulsoup4
lancedb
prettytable
pypdf
rapidocr-onnxruntime
google-generativeai
langchain_google_genai
python-multipart